+++ src/success/success.ts
@@ -2,14 +2,14 @@ import type { YobtaContext } from '../_types/YobtaContext.js'
 import type { YobtaSyncRule } from '../createRule/createRule.js'
 import { createRule } from '../createRule/createRule.js'
 
-interface Handler<I> {
+interface YobtaSuccessCallback<I> {
   (input: I, context: YobtaContext): void
 }
 
-export const successYobta = <I>(handle: Handler<I>): YobtaSyncRule<I, I> =>
+export const success = <I>(cb: YobtaSuccessCallback<I>): YobtaSyncRule<I, I> =>
   createRule((input: I, context) => {
     if (!context.errors.length) {
-      handle(input, context)
+      cb(input, context)
     }
     return input
   })